repaired bar back data

+++ src/main/java/com/beerandbeverages/backend/bar/dto/BarDetailsDto.java
@@ -1,7 +1,5 @@
 package com.beerandbeverages.backend.bar.dto;
 
-import java.math.BigDecimal;
-
 public record BarDetailsDto(
         String slug,
         Long cityId,
@@ -12,9 +10,10 @@ public record BarDetailsDto(
         String menuUrl,
         String siteUrl,
         String websiteUrl,
-        Integer priceLevel,
-        BigDecimal rating,
-        BigDecimal latitude,
-        BigDecimal longitude,
+        Byte priceLevel,
+        Double rating,
+        Double latitude,
+        Double longitude,
         Boolean isActive
 ) {}
+
