+++ src/main/java/com/beerandbeverages/backend/bar/dto/BarListItemDto.java
@@ -1,14 +1,13 @@
 package com.beerandbeverages.backend.bar.dto;
 
-import java.math.BigDecimal;
-
 public record BarListItemDto(
         String slug,
         Long cityId,
         String name,
         String address,
-        Integer priceLevel,
-        BigDecimal rating,
-        BigDecimal latitude,
-        BigDecimal longitude
+        Byte priceLevel,
+        Double rating,
+        Double latitude,
+        Double longitude
 ) {}
+
